Renamed/moved helper and utils/tests functions
