+++ tests/blog_list_api.test.js
@@ -3,7 +3,7 @@ const assert = require('node:assert')
 const supertest = require('supertest')
 const mongoose = require('mongoose')
 
-const helper = require('./test_helper')
+const helper = require('./blog_list_api_helper')
 const app = require('../app')
 const api = supertest(app)
