+++ tests/blog_list_functions.test.js
@@ -1,6 +1,6 @@
 const { test, describe } = require('node:test')
 const assert = require('node:assert')
-const { dummy, totalLikes, favoriteBlog, mostBlogs, mostLikes } = require('../utils/list_helper')
+const { dummy, totalLikes, favoriteBlog, mostBlogs, mostLikes } = require('../utils/blog_list_functions')
 const blogs = [
   {
     _id: '5a422a851b54a676234d17f7',
